Remove match case to make it compatible with earlier python versions

+++ elevenlabslib/helpers.py
@@ -8,20 +8,20 @@ default_headers = {'accept': '*/*'}
 def _api_call(requestType, path, headers, jsonData=None, filesData=None) -> requests.Response:
     if path[0] != "/":
         path = "/"+path
-    match requestType:
-        case "get":
-            response = requests.get(api_endpoint + path, headers=headers)
-        case "json":
-            response = requests.post(api_endpoint + path, headers=headers, json=jsonData)
-        case "del":
-            response = requests.delete(api_endpoint + path, headers=headers)
-        case "multipart":
-            if filesData is not None:
-                response = requests.post(api_endpoint + path, headers=headers, data=jsonData, files=filesData)
-            else:
-                response = requests.post(api_endpoint + path, headers=headers, data=jsonData)
-        case _:
-            raise Exception("Unknown API call type!")
+
+    if requestType == "get":
+        response = requests.get(api_endpoint + path, headers=headers)
+    elif requestType == "json":
+        response = requests.post(api_endpoint + path, headers=headers, json=jsonData)
+    elif requestType == "del":
+        response = requests.delete(api_endpoint + path, headers=headers)
+    elif requestType == "multipart":
+        if filesData is not None:
+            response = requests.post(api_endpoint + path, headers=headers, data=jsonData, files=filesData)
+        else:
+            response = requests.post(api_endpoint + path, headers=headers, data=jsonData)
+    else:
+        raise Exception("Unknown API call type!")
 
     if response.ok:
         return response
